+++ inst/NEWS.md
@@ -6,7 +6,16 @@ __and a Noise Component__
 =======================================================
 
 ### New Features, Improvements, Bug Fixes & Miscellaneous Edits
-* Fixed small bug for supplying univariate response data to `MoE_stepwise`.
+* Various fixes and improvements to initialisation when there expert network covariates:  
+    * Improved checks on the `resids` arg. to `MoE_mahala`.
+    * `MoE_control` arg. `exp.init$max.init` now defaults to `.Machine$integer.max`.
+    * Greatly expanded the `MoE_mahala` examples.
+* Fixed small `MoE_stepwise` bugs when  
+    * only one of `gating` or `expert` are supplied.
+    * univariate response `data` are supplied.
+* `noise_vol` now returns correction location for univariate data when `reciprocal=TRUE`.
+* `predict.MoEClust` now also returns the predicted values of the gating and expert networks.
+* `predict.MoEClust` now allows empty `newdata` for models with no covariates of any kind.
 
 ## MoEClust v1.2.4 - (_7<sup>th</sup> release [patch update]: 2019-12-11_)
 ### New Features, Improvements, Bug Fixes & Miscellaneous Edits